+++ src/collective/volto/formsupport/restapi/services/form_data/csv.py
@@ -1,14 +1,12 @@
-# -*- coding: utf-8 -*-
 from collective.volto.formsupport.interfaces import IFormDataStore
+from io import StringIO
 from plone.namedfile import NamedBlobFile
 from plone.restapi.serializer.converters import json_compatible
 from plone.restapi.services import Service
-from six import StringIO
 from zExceptions import NotFound
 from zope.component import getMultiAdapter
 
 import csv
-import six
 
 
 SKIP_ATTRS = ["block_id", "fields_labels", "fields_order"]