+++ src/collective/volto/formsupport/restapi/services/submit_form/post.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from collective.volto.formsupport import _
 from collective.volto.formsupport.interfaces import ICaptchaSupport
 from collective.volto.formsupport.interfaces import IFormDataStore
@@ -17,7 +16,6 @@ from plone.registry.interfaces import IRegistry
 from plone.restapi.deserializer import json_body
 from plone.restapi.services import Service
 from plone.schema.email import _isemail
-from Products.CMFPlone.interfaces.controlpanel import IMailSchema
 from xml.etree.ElementTree import Element
 from xml.etree.ElementTree import ElementTree
 from xml.etree.ElementTree import SubElement
@@ -34,7 +32,6 @@ import codecs
 import logging
 import math
 import os
-import six
 
 
 logger = logging.getLogger(__name__)