+++ src/collective/volto/formsupport/restapi/services/submit_form/post.py
@@ -277,6 +277,8 @@ class SubmitPost(Service):
             "parameters": self.filter_parameters(),
             "url": self.context.absolute_url(),
             "title": self.context.Title(),
+            "mail_header": self.block.get("mail_header", None),
+            "mail_footer": self.block.get("mail_footer", None),
         }
         return message_template(**parameters)